Move windows files to windows folder
