+++ src/windows/help_window.c
@@ -1,4 +1,4 @@
-#include "help_window/help_window.h"
+#include "windows/help_window.h"
 #include "utils.h"
 
 #include <stdlib.h>
