+++ src/windows/input_window.c
@@ -1,4 +1,4 @@
-#include "input_window/input_window.h"
+#include "windows/input_window.h"
 #include "utils.h"
 #include <stdlib.h>
 #include <ctype.h>
